requirements.txt
bentoml==1.0.23
click==8.1.3
joblib==1.3.1
lxml==4.9.3
matplotlib==3.7.1
mlflow==2.2.2
numpy==1.24.4
pandas==1.5.3
Pillow==10.0.0
Pillow==10.0.0
python-dotenv==1.0.0
rich==13.4.2
scikit_learn==1.2.0
scipy==1.11.1
SQLAlchemy==1.4.41
statsmodels==0.14.0
typing_extensions==4.7.1
zenml==0.42.1